Dropped the Builder thingy. Didn't make enough sense.

+++ Bullet-Inferno/src/se/dat255/bulletinferno/model/Destructible.java
@@ -7,7 +7,7 @@ public interface Destructible {
 	 * @param damage
 	 *            The damage to be dealt.
 	 */
-	public void takeDamage(int damage);
+	public void takeDamage(float damage);
 
 	/**
 	 * Returns the object's current health.
